feat: move pages to /profile
